chore: 총개수 구하는 서브쿼리 개별 쿼리로 나누기 전 코드 보존용 커밋
- 기존에는 한쿼리에 총개수와 데이터를 가져오는 서브쿼리를 한번에 처리했는데, 변경 예정
- 이유는 총 개수를 구하는 서브 쿼리만 따로 수행하면 실제 데이터 페이지 접근 없이 인덱스 트리만 탐색하여 개수를 계산해 효율적일 수 있기 때문

+++ blog-springboot-jpa/src/main/java/com/yhs/blog/springboot/jpa/domain/file/service/infrastructure/s3/impl/S3ServiceImpl.java
@@ -223,6 +223,10 @@ public class S3ServiceImpl implements S3Service {
     @Override
     public CompletableFuture<Void> processDeletePostS3Operation(List<String> toBeDeletedFileUrls, String blogId) {
 
+        if (toBeDeletedFileUrls == null || toBeDeletedFileUrls.isEmpty()) {
+            return CompletableFuture.completedFuture(null);
+        }
+
         try {
             for (String toBeDeletedFileUrl : toBeDeletedFileUrls) {
                 deleteFile(toBeDeletedFileUrl, blogId);
@@ -230,7 +234,7 @@ public class S3ServiceImpl implements S3Service {
 
         } catch (Exception e) {
             log.error("[S3ServiceImpl] processDeletePostS3Operation() - AwSS3 실패 에러", e);
-            
+
         }
 
         return CompletableFuture.completedFuture(null);
